Update IVaultService to use SecureLargeBuffer
- Updated IVaultService to use SecureLargeBuffer instead of raw byte[] array to increase security
- Updated IVaultService to expect ReadOnlySpan instead of byte[] array to prevent accidentally leaking data when GC moves underlying memory
- Updated affected tests and improved them slightly by utilizing try/finally and using blocks

+++ VaultCrypt.Tests/Services/VaultServiceTests.cs
@@ -33,17 +33,22 @@ namespace VaultCrypt.Tests.Services
             byte[] password = RandomNumberGenerator.GetBytes(32);
             int iterations = 10;
 
-            _service.CreateVault(path, fileName, password, iterations);
-
-            using (FileStream fs = new FileStream($"{path}\\{fileName}.vlt", FileMode.Open, FileAccess.Read))
+            try
             {
-                Assert.Equal(_session.VAULT_READER.HeaderSize, fs.Length);
-                Assert.Equal(VaultSession.NewestVaultVersion, fs.ReadByte());
-                Assert.False(_session.VAULT_READER.ReadSalt(fs).AsSpan.SequenceEqual(new byte[_session.VAULT_READER.SaltSize])); //Asserting that salt is not empty (zeroed out value)
-                Assert.Equal(iterations, _session.VAULT_READER.ReadIterationsNumber(fs));
-            }
+                _service.CreateVault(path, fileName, password, iterations);
 
-            File.Delete($"{path}\\{fileName}.vlt");
+                using (FileStream fs = new FileStream($"{path}\\{fileName}.vlt", FileMode.Open, FileAccess.Read))
+                {
+                    Assert.Equal(_session.VAULT_READER.HeaderSize, fs.Length);
+                    Assert.Equal(VaultSession.NewestVaultVersion, fs.ReadByte());
+                    Assert.False(_session.VAULT_READER.ReadSalt(fs).AsSpan.SequenceEqual(new byte[_session.VAULT_READER.SaltSize])); //Asserting that salt is not empty (zeroed out value)
+                    Assert.Equal(iterations, _session.VAULT_READER.ReadIterationsNumber(fs));
+                }
+            }
+            finally
+            {
+                File.Delete($"{path}\\{fileName}.vlt");
+            }
         }
 
         [Fact]
@@ -54,11 +59,16 @@ namespace VaultCrypt.Tests.Services
             byte[] password = RandomNumberGenerator.GetBytes(32);
             int iterations = 10;
 
-            _service.CreateVault(path, fileName, password, iterations);
+            try
+            {
+                _service.CreateVault(path, fileName, password, iterations);
 
-            Assert.True(File.Exists($"{path}\\{fileName}.vlt"));
-
-            Directory.Delete(path!, recursive: true);
+                Assert.True(File.Exists($"{path}\\{fileName}.vlt"));
+            }
+            finally
+            {
+                Directory.Delete(path!, recursive: true);
+            }
         }
 
         [Fact]
@@ -67,7 +77,6 @@ namespace VaultCrypt.Tests.Services
             string nonNullValue = "TEXT";
             Assert.Throws<ArgumentNullException>(() => _service.CreateVault(null!, nonNullValue, new byte[1], 1));
             Assert.Throws<ArgumentNullException>(() => _service.CreateVault(NormalizedPath.From(nonNullValue), null!, new byte[1], 1));
-            Assert.Throws<ArgumentNullException>(() => _service.CreateVault(NormalizedPath.From(nonNullValue), nonNullValue, null!, 1));
         }
 
         [Fact]
@@ -101,22 +110,28 @@ namespace VaultCrypt.Tests.Services
         [Fact]
         void CreateSessionFromFileSetsValuesCorrectly()
         {
-            var vaultFile = TestsHelper.CreateVaultFile();
 
-            _service.CreateSessionFromFile(new byte[16], vaultFile);
+            NormalizedPath vaultFile = null!;
+            try
+            {
+                vaultFile = TestsHelper.CreateVaultFile();
+                _service.CreateSessionFromFile(new byte[16], vaultFile);
 
-            Assert.Equal(PasswordHelper.DeriveKey(new byte[16], new byte[32], 1000), _session.KEY);
-            Assert.Equal(vaultFile, _session.VAULTPATH);
-            Assert.Empty(_session.ENCRYPTED_FILES);
-            Assert.True(_session.VAULT_READER is VaultV0Reader);
 
-            File.Delete(vaultFile!);
+                Assert.True(TestsHelper.CreateKey(new byte[16], new byte[32], 1000).SequenceEqual(_session.KEY.AsSpan[..PasswordHelper.KeySize]));
+                Assert.Equal(vaultFile, _session.VAULTPATH);
+                Assert.Empty(_session.ENCRYPTED_FILES);
+                Assert.True(_session.VAULT_READER is VaultV0Reader);
+            }
+            finally
+            {
+                if (vaultFile is not null) File.Delete(vaultFile!);
+            }
         }
 
         [Fact]
         void CreateSessionFromFileThrowsForNullValues()
         {
-            Assert.Throws<ArgumentNullException>(() => _service.CreateSessionFromFile(null!, NormalizedPath.From("VALID")));
             Assert.Throws<ArgumentNullException>(() => _service.CreateSessionFromFile(new byte[1], null!));
         }
 
@@ -133,20 +148,32 @@ namespace VaultCrypt.Tests.Services
         {
             _session.ENCRYPTED_FILES.Clear();
             byte expectedFileListCount = (byte)RandomNumberGenerator.GetInt32(10);
-            var vaultInformation = TestsHelper.CreateVaultFileWithEncryptedFileList(expectedFileListCount, _session);
 
-            using (FileStream fs = new FileStream(vaultInformation.Item1!, FileMode.Open, FileAccess.Read))
+            (NormalizedPath, EncryptionOptions.FileEncryptionOptions[]) vaultInformation = (null!, null!);
+            try
             {
-                _service.RefreshEncryptedFilesList(fs);
+                vaultInformation = TestsHelper.CreateVaultFileWithEncryptedFileList(expectedFileListCount, _session);
+
+                using (FileStream fs = new FileStream(vaultInformation.Item1!, FileMode.Open, FileAccess.Read))
+                {
+                    _service.RefreshEncryptedFilesList(fs);
+                }
+                var fileList = _session.ENCRYPTED_FILES.ToList();
+                Assert.Equal(expectedFileListCount, _session.ENCRYPTED_FILES.Count);
+                for (int i = 0; i < expectedFileListCount; i++)
+                {
+                    Assert.Equal(Encoding.UTF8.GetString(vaultInformation.Item2[i].FileName.AsSpan), fileList[i].Value.FileName);
+                    Assert.Equal(EncryptionAlgorithm.GetEncryptionAlgorithmInfo[vaultInformation.Item2[i].EncryptionAlgorithm].Name, fileList[i].Value.EncryptionAlgorithm);
+                }
             }
-            var fileList = _session.ENCRYPTED_FILES.ToList();
-            Assert.Equal(expectedFileListCount, _session.ENCRYPTED_FILES.Count);
-            for (int i = 0; i < expectedFileListCount; i++)
+            finally
             {
-                Assert.Equal(Encoding.UTF8.GetString(vaultInformation.Item2[i].FileName), fileList[i].Value.FileName);
-                Assert.Equal(EncryptionAlgorithm.GetEncryptionAlgorithmInfo[vaultInformation.Item2[i].EncryptionAlgorithm].Name, fileList[i].Value.EncryptionAlgorithm);
+                if (vaultInformation.Item1 is not null) File.Delete(vaultInformation.Item1!);
+                foreach (var item in vaultInformation.Item2)
+                {
+                    item.Dispose();
+                }
             }
-            File.Delete(vaultInformation.Item1!);
         }
 
         [Fact]
@@ -159,39 +186,73 @@ namespace VaultCrypt.Tests.Services
         void TrimVaultTrimsVaultAndSavesItToAFile()
         {
             byte[] password = Encoding.UTF8.GetBytes("TrimVaultTrims");
-            (NormalizedPath, EncryptionOptions.FileEncryptionOptions[]) tuple = TestsHelper.CreateVaultFileWithEncryptedFileList(10, _session, password);
-            long oldVaultSize = new FileInfo(tuple.Item1!).Length;
-            //Trim vault uses vault provided at VaultSession.VAULTPATH where VaultSession is the one provided when creating service
-            using (FileStream fs = new FileStream(tuple.Item1.Value, FileMode.Open, FileAccess.Read))
+
+            (NormalizedPath, EncryptionOptions.FileEncryptionOptions[]) vaultInformation = (null!, null!);
+            NormalizedPath newVaultPath = null!;
+            try
             {
-                TestsHelper.SetVaultSessionFromStream(_session, fs, password);
+                vaultInformation = TestsHelper.CreateVaultFileWithEncryptedFileList(numberOfFiles: 10, _session, password);
+                
+                long oldVaultSize = new FileInfo(vaultInformation.Item1!).Length;
+                //Trim vault uses vault provided at VaultSession.VAULTPATH where VaultSession is the one provided when creating service
+                using (FileStream fs = new FileStream(vaultInformation.Item1.Value, FileMode.Open, FileAccess.Read))
+                {
+                    TestsHelper.SetVaultSessionFromStream(_session, fs, password);
+                }
+                _session.ENCRYPTED_FILES.Remove(_session.ENCRYPTED_FILES.First().Key);
+
+                try
+                {
+
+                    _service.TrimVault(new ProgressionContext());
+                    newVaultPath = NormalizedPath.From(vaultInformation.Item1!.Value[..^4] + "_TRIMMED.vlt");
+                    long newVaultSize = new FileInfo(newVaultPath!).Length;
+                    Assert.True(newVaultSize < oldVaultSize);
+                }
+                finally
+                {
+                    if (newVaultPath is not null) File.Delete(newVaultPath!);
+                }
             }
-            _session.ENCRYPTED_FILES.Remove(_session.ENCRYPTED_FILES.First().Key);
-            _service.TrimVault(new ProgressionContext());
-            long newVaultSize = new FileInfo((tuple.Item1!.Value[..^4] + "_TRIMMED.vlt")).Length;
-
-            Assert.True(newVaultSize < oldVaultSize);
-
-            File.Delete(tuple.Item1!);
-            File.Delete((tuple.Item1!.Value[..^4] + "_TRIMMED.vlt"));
+            finally
+            {
+                if (vaultInformation.Item1 is not null) File.Delete(vaultInformation.Item1!);
+                
+            }
+            
         }
 
         [Fact]
         void TrimVaultReturnsSameVaultIfThereIsNothingToTrim()
         {
             byte[] password = Encoding.UTF8.GetBytes("TrimVaultReturnsSameVaultIfThereIsNothingToTrim");
-            (NormalizedPath, EncryptionOptions.FileEncryptionOptions[]) tuple = TestsHelper.CreateVaultFileWithEncryptedFileList(10, _session, password);
-            //Trim vault uses vault provided at VaultSession.VAULTPATH where VaultSession is the one provided when creating service
-            using (FileStream fs = new FileStream(tuple.Item1.Value, FileMode.Open, FileAccess.Read))
+
+
+            (NormalizedPath, EncryptionOptions.FileEncryptionOptions[]) vaultInformation = (null!, null!);
+            NormalizedPath newVaultPath = null!;
+            try
             {
-                TestsHelper.SetVaultSessionFromStream(_session, fs, password);
+                vaultInformation = TestsHelper.CreateVaultFileWithEncryptedFileList(10, _session, password);
+                //Trim vault uses vault provided at VaultSession.VAULTPATH where VaultSession is the one provided when creating service
+                using (FileStream fs = new FileStream(vaultInformation.Item1.Value, FileMode.Open, FileAccess.Read))
+                {
+                    TestsHelper.SetVaultSessionFromStream(_session, fs, password);
+                }
+                try
+                {
+                    _service.TrimVault(new ProgressionContext());
+                    newVaultPath = NormalizedPath.From(vaultInformation.Item1!.Value[..^4] + "_TRIMMED.vlt");
+                    Assert.Equal(new FileInfo(vaultInformation.Item1!).Length, new FileInfo(newVaultPath!).Length);
+                }
+                finally
+                {
+                    File.Delete(newVaultPath!);
+                }
             }
-            _service.TrimVault(new ProgressionContext());
-
-            Assert.Equal(new FileInfo(tuple.Item1!).Length, new FileInfo((tuple.Item1!.Value[..^4] + "_TRIMMED.vlt")).Length);
-
-            File.Delete(tuple.Item1!);
-            File.Delete((tuple.Item1!.Value[..^4] + "_TRIMMED.vlt"));
+            finally
+            {
+                if (vaultInformation.Item1 is not null) File.Delete(vaultInformation.Item1!);
+            }
         }
 
         [Fact]
@@ -231,6 +292,24 @@ namespace VaultCrypt.Tests.Services
 
             Assert.True(newVaultSize < oldVaultSize);
 
+            File.Delete(tuple.Item1!);
+        }
+
+        [Fact]
+        void DeleteFileFromVaultChangesEncryptedFileListCount()
+        {
+            (NormalizedPath, EncryptionOptions.FileEncryptionOptions[]) tuple = TestsHelper.CreateVaultFileWithEncryptedFileList(10, _session);
+            long oldVaultSize = new FileInfo(tuple.Item1!).Length;
+
+            var oldFileListCount = _session.ENCRYPTED_FILES.Count;
+            _service.DeleteFileFromVault(_session.ENCRYPTED_FILES.Last().Key, new ProgressionContext());
+            using (FileStream vaultFS = new FileStream(_session.VAULTPATH!, FileMode.Open, FileAccess.Read))
+            {
+                _service.RefreshEncryptedFilesList(vaultFS);
+            }
+            var actualFileListCount = _session.ENCRYPTED_FILES.Count;
+
+            Assert.Equal(oldFileListCount - 1, actualFileListCount);
             File.Delete(tuple.Item1!);
         }
